+++ common/collections/collections.go
@@ -19,10 +19,3 @@ package collections
 type Grouper interface {
 	Group(key interface{}, items interface{}) (interface{}, error)
 }
-
-// Slicer definse a very generic way to create a typed slice. This is used
-// in collections.Slice template func to get types such as Pages, PageGroups etc.
-// instead of the less useful []interface{}.
-type Slicer interface {
-	Slice(items interface{}) (interface{}, error)
-}
